backend-app/requirements.txt
fastapi
uvicorn
google-generativeai
chromadb
sentence-transformers
sqlalchemy[sqlite]
psycopg2-binary